requirements.txt
streamlit==1.32.0
mysql-connector-python==8.3.0
pandas==2.2.0
python-dotenv==1.0.1 